refactor: decompose app/index.tsx and add deferred useMyActiveRoom test
app/index.tsx was 954 lines after feature 018 (multiplayer room
management) landed. Split it into:
- utils/homeStats.ts: getTotalDrinks, getTopDrinker (+ tests)
- components/home/: HomeSplash, CurrentGameCard, HistoryStatsCard,
  CancelGameModal, JoinRoomModal (the latter two extracted in a
  second pass once the first split still left index.tsx at 527 lines,
  over the 400-line ceiling)
- hooks/useHomeRoomActions.ts: composes useHostRoomCreate,
  useMyActiveRoom, useRegisteredRoomJoin, useGuestRoomJoin,
  useRoomExit plus their handlers and modal state

app/index.tsx is now 370 lines of layout/composition. Pure move
against the freshly-merged 018 code, no behavior edits — the
pre-existing __tests__/app/index.platform.test.tsx passes unmodified.
Two apostrophes in the extracted JoinRoomModal needed HTML-escaping
(You're -> You&apos;re) since the app/index.tsx-specific
react/no-unescaped-entities override in eslint.config.js no longer
applies once that JSX moved to a new file; rendered text is unchanged.

Also adds the deferred __tests__/hooks/useMyActiveRoom.test.ts
(6 tests), mirroring the post-018 useRoomLobby.test.ts pattern —
its mock seam (utils/supabaseClient) and reference test were both
in-flight on 153 until now.
